+++ src/transports/sentry_transport_curl.c
@@ -17,7 +17,7 @@ typedef struct curl_transport_state_s {
     sentry_dsn_t *dsn;
     CURL *curl_handle;
     char *user_agent;
-    char *http_proxy;
+    char *proxy;
     char *ca_certs;
     sentry_rate_limiter_t *ratelimiter;
     bool debug;
@@ -54,7 +54,7 @@ sentry__curl_bgworker_state_free(void *_state)
     sentry__rate_limiter_free(state->ratelimiter);
     sentry_free(state->ca_certs);
     sentry_free(state->user_agent);
-    sentry_free(state->http_proxy);
+    sentry_free(state->proxy);
     sentry_free(state);
 }
 
@@ -85,7 +85,7 @@ sentry__curl_transport_start(
         };
 
         if (!sentry__check_min_version(
-                curl_version, (sentry_version_t) { 7, 10, 7 })) {
+                curl_version, (sentry_version_t) { 7, 21, 7 })) {
             SENTRY_WARNF("`libcurl` is at unsupported version `%u.%u.%u`",
                 curl_version.major, curl_version.minor, curl_version.patch);
             return 1;
@@ -101,7 +101,7 @@ sentry__curl_transport_start(
     curl_bgworker_state_t *state = sentry__bgworker_get_state(bgworker);
 
     state->dsn = sentry__dsn_incref(options->dsn);
-    state->http_proxy = sentry__string_clone(options->http_proxy);
+    state->proxy = sentry__string_clone(options->proxy);
     state->user_agent = sentry__string_clone(options->user_agent);
     state->ca_certs = sentry__string_clone(options->ca_certs);
     state->curl_handle = curl_easy_init();
@@ -215,8 +215,8 @@ sentry__curl_send_task(void *_envelope, void *_state)
     curl_easy_setopt(curl, CURLOPT_HEADERDATA, (void *)&info);
     curl_easy_setopt(curl, CURLOPT_HEADERFUNCTION, header_callback);
 
-    if (state->http_proxy) {
-        curl_easy_setopt(curl, CURLOPT_PROXY, state->http_proxy);
+    if (state->proxy) {
+        curl_easy_setopt(curl, CURLOPT_PROXY, state->proxy);
     }
     if (state->ca_certs) {
         curl_easy_setopt(curl, CURLOPT_CAINFO, state->ca_certs);